requirements.txt
beautifulsoup4==4.6.3
cachetools==2.1.0
certifi==2018.10.15
chardet==3.0.4
clearbit==0.1.7
Click==7.0
decorator==4.3.0
dnspython==1.15.0
google-api-core==1.6.0a1
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.6.0
google-cloud-bigquery-datatransfer==0.1.1
google-cloud-core==0.28.1
google-cloud-language==1.1.0
google-cloud-pubsub==0.38.0
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
grpc-google-iam-v1==0.11.4
grpcio==1.16.0
httplib2==0.11.3
idna==2.7
protobuf==3.6.1
psutil==5.4.7
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.5
pytz==2018.6
requests==2.20.0
rsa==4.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.24
validators==0.12.2